Don't allow uninvoice client invoice when claim is attached ALLY-1424

+++ app/Http/Controllers/Admin/ClientInvoiceController.php
@@ -112,6 +112,10 @@ class ClientInvoiceController extends Controller
             return new ErrorResponse(400, "This invoice cannot be removed because it has payments assigned.");
         }
 
+        if (filled($invoice->claimInvoice)) {
+            return new ErrorResponse(400, "This invoice cannot be removed because it has a Claim attached.");
+        }
+
         if ($invoice->delete()) {
             return new SuccessResponse("The invoice has been removed.");
         }
